Completed ALL tests, fixed a dependency build file
