+++ src/lambdas/FileCoordinator/test/index.test.ts
@@ -1,46 +1,48 @@
-import {describe, expect, test, jest, beforeEach} from '@jest/globals'
-import * as DynamoDB from '../../../lib/vendor/AWS/DynamoDB'
-import * as StepFunctions from '../../../lib/vendor/AWS/StepFunctions'
-import * as chai from 'chai'
-import {getFixture, testContext} from '../../../util/mocha-setup'
-import {handler} from '../src/index'
-import {ScheduledEvent} from 'aws-lambda'
-import {UnexpectedError} from '../../../util/errors'
-const expect = chai.expect
-import path from 'path'
-import {fileURLToPath} from 'url'
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
-const localFixture = getFixture.bind(null, __dirname)
+import {testContext} from '../../../util/jest-setup'
+import {describe, expect, test, jest} from '@jest/globals'
+const {default: eventMock} = await import('./fixtures/ScheduledEvent.json', {assert: {type: 'json'}})
+
+const scanMock = jest.fn()
+jest.unstable_mockModule('../../../lib/vendor/AWS/DynamoDB', () => ({
+  scan: scanMock,
+  deleteItem: jest.fn(),
+  query: jest.fn(),
+  updateItem: jest.fn()
+}))
+const startExecutionMock = jest.fn()
+jest.unstable_mockModule('../../../lib/vendor/AWS/StepFunctions', () => ({
+  startExecution: startExecutionMock
+}))
+
+const {handler} = await import('./../src')
 
 describe('#FileCoordinator', () => {
   process.env.DynamoDBTableFiles = 'Files'
   const context = testContext
-  const event = localFixture('ScheduledEvent.json') as ScheduledEvent
-  let scanStub: sinon.SinonStub
-  let startExecutionStub: sinon.SinonStub
-  beforeEach(() => {
-    scanStub = sinon.stub(DynamoDB, 'scan')
-    startExecutionStub = sinon.stub(StepFunctions, 'startExecution')
-  })
+  const event = JSON.parse(JSON.stringify(eventMock))
   test('should handle scheduled event (with no events)', async () => {
-    scanStub.returns(localFixture('scan-204-NoContent.json'))
-    startExecutionStub.returns(localFixture('startExecution-200-OK.json'))
+    const {default: scanResponse} = await import('./fixtures/scan-204-NoContent.json', {assert: {type: 'json'}})
+    const {default: startExecutionResponse} = await import('./fixtures/startExecution-200-OK.json', {assert: {type: 'json'}})
+    scanMock.mockReturnValue(scanResponse)
+    startExecutionMock.mockReturnValue(startExecutionResponse)
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(200)
-    expect(startExecutionStub.callCount).to.equal(0)
+    expect(output.statusCode).toEqual(200)
+    expect(startExecutionMock).toHaveBeenCalledTimes(0)
   })
   test('should handle scheduled event (with 1 event)', async () => {
-    scanStub.returns(localFixture('scan-200-OK.json'))
-    startExecutionStub.returns(localFixture('startExecution-200-OK.json'))
+    const {default: scanResponse} = await import('./fixtures/scan-200-OK.json', {assert: {type: 'json'}})
+    const {default: startExecutionResponse} = await import('./fixtures/startExecution-200-OK.json', {assert: {type: 'json'}})
+    scanMock.mockReturnValue(scanResponse)
+    startExecutionMock.mockReturnValue(startExecutionResponse)
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(200)
-    expect(startExecutionStub.callCount).to.equal(1)
+    expect(output.statusCode).toEqual(200)
+    expect(startExecutionMock).toHaveBeenCalled()
   })
   describe('#AWSFailure', () => {
     test('AWS.DynamoDB.DocumentClient.scan', async () => {
-      scanStub.returns(undefined)
-      expect(handler(event, context)).to.be.rejectedWith(UnexpectedError)
+      const message = 'AWS request failed'
+      scanMock.mockReturnValue(undefined)
+      await expect(handler(event, context)).rejects.toThrowError(message)
     })
   })
 })
