+++ src/lambdas/WebhookFeedly/test/index.test.ts
@@ -1,107 +1,120 @@
-import * as sinon from 'sinon'
-import * as DynamoDB from '../../../lib/vendor/AWS/DynamoDB'
-import * as StepFunctions from '../../../lib/vendor/AWS/StepFunctions'
-import * as SQS from '../../../lib/vendor/AWS/SQS'
-import {getFixture, testContext} from '../../../util/mocha-setup'
-import * as chai from 'chai'
-import {handler} from '../src/index'
-import {APIGatewayEvent} from 'aws-lambda'
+import {describe, expect, test, jest, beforeEach} from '@jest/globals'
+import {testContext} from '../../../util/jest-setup'
 import {v4 as uuidv4} from 'uuid'
-import {UnexpectedError} from '../../../util/errors'
-const expect = chai.expect
-import path from 'path'
-import {fileURLToPath} from 'url'
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
-const localFixture = getFixture.bind(null, __dirname)
+import {CustomAPIGatewayRequestAuthorizerEvent} from '../../../types/main'
 const fakeUserId = uuidv4()
+
+const queryMock = jest.fn()
+const updateItem = jest.fn().mockReturnValue({})
+const deleteItemMock = jest.fn().mockReturnValue({})
+jest.unstable_mockModule('../../../lib/vendor/AWS/DynamoDB', () => ({
+  updateItem: updateItem,
+  deleteItem: deleteItemMock,
+  query: queryMock,
+  scan: jest.fn()
+}))
+
+jest.unstable_mockModule('../../../lib/vendor/AWS/SQS', () => ({
+  sendMessage: jest.fn().mockReturnValue({
+    MD5OfMessageBody: '44dd2fc26e4186dc12b8e67ccb9a9435',
+    MD5OfMessageAttributes: 'e95833d661f4007f9575877843f475ed',
+    MessageId: 'e990c66f-23f6-4982-9274-a5a533ceb6dc'
+  }),
+  subscribe: jest.fn()
+}))
+
+jest.unstable_mockModule('../../../lib/vendor/AWS/StepFunctions', () => ({
+  startExecution: jest.fn().mockReturnValue({
+    executionArn: 'arn:aws:states:us-west-2:203465012143:execution:MultipartUpload:1666060419059',
+    startDate: new Date()
+  })
+}))
+
+const {default: handleFeedlyEventResponse} = await import('./fixtures/handleFeedlyEvent-200-OK.json', {assert: {type: 'json'}})
+const {default: queryNoContentResponse} = await import('./fixtures/query-204-NoContent.json', {assert: {type: 'json'}})
+const {default: querySuccessResponse} = await import('./fixtures/query-200-OK.json', {assert: {type: 'json'}})
+
+const {default: eventMock} = await import('./fixtures/APIGatewayEvent.json', {assert: {type: 'json'}})
+const {handler} = await import('./../src')
 
 describe('#WebhookFeedly', () => {
   const context = testContext
-  let event: APIGatewayEvent
+  let event: CustomAPIGatewayRequestAuthorizerEvent
   beforeEach(() => {
-    event = localFixture('APIGatewayEvent.json') as APIGatewayEvent
+    event = JSON.parse(JSON.stringify(eventMock))
   })
-  afterEach(() => {
-    sinon.restore()
-  })
-  it('should trigger the download of a new file immediately (if not present)', async () => {
+  test('should trigger the download of a new file immediately (if not present)', async () => {
     event.requestContext.authorizer!.principalId = fakeUserId
-    event.body = JSON.stringify(localFixture('handleFeedlyEvent-200-OK.json'))
-    sinon.stub(DynamoDB, 'updateItem').resolves({})
-    sinon.stub(DynamoDB, 'query').resolves(localFixture('query-204-NoContent.json'))
-    sinon.stub(StepFunctions, 'startExecution').resolves({
-      executionArn: 'arn:aws:states:us-west-2:203465012143:execution:MultipartUpload:1666060419059',
-      startDate: new Date()
-    })
+    event.body = JSON.stringify(handleFeedlyEventResponse)
+    queryMock.mockReturnValue(queryNoContentResponse)
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(202)
+    expect(output.statusCode).toEqual(202)
     const body = JSON.parse(output.body)
-    expect(body.body.status).to.equal('Initiated')
+    expect(body.body.status).toEqual('Initiated')
   })
-  it('should trigger the download of a new file later (if not present)', async () => {
+  test('should trigger the download of a new file later (if not present)', async () => {
     event.requestContext.authorizer!.principalId = fakeUserId
-    const eventBody = localFixture('handleFeedlyEvent-200-OK.json')
+    const eventBody = handleFeedlyEventResponse
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     eventBody.backgroundMode = 1
     event.body = JSON.stringify(eventBody)
-    sinon.stub(DynamoDB, 'updateItem').resolves({})
-    sinon.stub(DynamoDB, 'query').resolves(localFixture('query-204-NoContent.json'))
+    queryMock.mockReturnValue(queryNoContentResponse)
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(202)
+    expect(output.statusCode).toEqual(202)
     const body = JSON.parse(output.body)
-    expect(body.body.status).to.equal('Accepted')
+    expect(body.body.status).toEqual('Accepted')
   })
-  it('should dispatch a message to the users device (if the file already exists)', async () => {
+  test('should dispatch a message to the users device (if the file already exists)', async () => {
     event.requestContext.authorizer!.principalId = fakeUserId
-    sinon.stub(DynamoDB, 'updateItem').resolves({})
-    sinon.stub(DynamoDB, 'query').resolves(localFixture('query-200-OK.json'))
-    sinon.stub(SQS, 'sendMessage').resolves({
-      MD5OfMessageBody: '44dd2fc26e4186dc12b8e67ccb9a9435',
-      MD5OfMessageAttributes: 'e95833d661f4007f9575877843f475ed',
-      MessageId: 'e990c66f-23f6-4982-9274-a5a533ceb6dc'
+    queryMock.mockReturnValue(querySuccessResponse)
+    event.body = JSON.stringify(handleFeedlyEventResponse)
+    const output = await handler(event, context)
+    expect(output.statusCode).toEqual(200)
+    const body = JSON.parse(output.body)
+    expect(body.body.status).toEqual('Dispatched')
+  })
+  test('should fail gracefully if the startExecution fails', async () => {
+    event.requestContext.authorizer!.principalId = fakeUserId
+    event.body = JSON.stringify(handleFeedlyEventResponse)
+    updateItem.mockImplementation(() => {
+      throw new Error('Update failed')
     })
-    event.body = JSON.stringify(localFixture('handleFeedlyEvent-200-OK.json'))
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(200)
-    const body = JSON.parse(output.body)
-    expect(body.body.status).to.equal('Dispatched')
+    expect(output.statusCode).toEqual(500)
   })
-  it('should fail gracefully if the startExecution fails', async () => {
-    event.requestContext.authorizer!.principalId = fakeUserId
-    event.body = JSON.stringify(localFixture('handleFeedlyEvent-200-OK.json'))
-    sinon.stub(DynamoDB, 'updateItem').rejects('Error')
-    const output = await handler(event, context)
-    expect(output.statusCode).to.equal(500)
-  })
-  it('should handle an invalid request body', async () => {
+  test('should handle an invalid request body', async () => {
     event.body = JSON.stringify({})
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(400)
+    expect(output.statusCode).toEqual(400)
     const body = JSON.parse(output.body)
-    expect(body.error.message).to.have.property('articleURL')
-    expect(body.error.message.articleURL[0]).to.have.string('is required')
+    expect(body.error.message).toHaveProperty('articleURL')
+    expect(body.error.message.articleURL[0]).toEqual('articleURL is required')
   })
-  it('should handle a missing user ID', async () => {
+  test('should handle a missing user ID', async () => {
     event.requestContext.authorizer!.principalId = 'unknown'
-    event.body = JSON.stringify(localFixture('handleFeedlyEvent-200-OK.json'))
-    expect(handler(event, context)).to.be.rejectedWith(UnexpectedError)
+    event.body = JSON.stringify(handleFeedlyEventResponse)
+    const output = await handler(event, context)
+    expect(output.statusCode).toEqual(500)
+    const body = JSON.parse(output.body)
+    expect(body.error.code).toEqual('custom-5XX-generic')
+    expect(body.error.message).toEqual('AWS request failed')
   })
-  it('should handle an invalid event body', async () => {
+  test('should handle an invalid event body', async () => {
     event.body = 'hello'
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(400)
+    expect(output.statusCode).toEqual(400)
     const body = JSON.parse(output.body)
-    expect(body.error.code).to.equal('custom-4XX-generic')
-    expect(body.error.message).to.equal('Request body must be valid JSON')
+    expect(body.error.code).toEqual('custom-4XX-generic')
+    expect(body.error.message).toEqual('Request body must be valid JSON')
   })
-  it('should handle an invalid (non-YouTube) URL', async () => {
-    event.body = JSON.stringify(localFixture('handleFeedlyEvent-400-InvalidURL.json'))
+  test('should handle an invalid (non-YouTube) URL', async () => {
+    const {default: handleFeedlyEventInvalidResponse} = await import('./fixtures/handleFeedlyEvent-400-InvalidURL.json', {assert: {type: 'json'}})
+    event.body = JSON.stringify(handleFeedlyEventInvalidResponse)
     const output = await handler(event, context)
-    expect(output.statusCode).to.equal(400)
+    expect(output.statusCode).toEqual(400)
     const body = JSON.parse(output.body)
-    expect(body.error.message).to.have.property('articleURL')
-    expect(body.error.message.articleURL[0]).to.have.string('not a valid YouTube URL')
+    expect(body.error.message).toHaveProperty('articleURL')
+    expect(body.error.message.articleURL[0]).toEqual('is not a valid YouTube URL')
   })
 })
